+++ test/baf/zero_degree_b_spline_basis_function_test.cc
@@ -23,7 +23,7 @@ using namespace splinelib::src;
 
 class KnotVectorMock000111 : public baf::KnotVector {
  public:
-  ParametricCoordinate GetKnot(size_t knot_num) const override {
+  ParametricCoordinate GetKnot(int knot_num) const override {
     return knts[knot_num];
   }
